+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## ??? - 2023-08-31
+
+### Added
+
+- new db collection for history of study rules, new methods to get, delete and add study rules. 
+- new uploadedAt index model for study rules 
+- new gRPC endpoints: 
+  - `GetStudyRulesHistory`: get study rules history for specified study key that fullfill query criteria with pagination and sorting options 
+  - `GetCurrentStudyRules`: get current study rules for specified study key 
+  - `RemoveStudyRulesVersion`: deletes study rules version with specified id.
+
+### Changed
+
+- gRPC endpoints:
+  - `DeleteStudy`: study rules history is also deleted 
+  - `SaveStudyRules`: new study rules are added in study rules collection 
+
+
 ## ??? - 2023-08-01
 
 ### Added